requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
fuzzywuzzy==0.17.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
pycryptodome==3.9.0
python-dateutil==2.8.0
python-Levenshtein-wheels==0.13.1
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.6
